Applying JWT Security in supplier_service(delete suppliers)

+++ js/supplier_service.js
@@ -231,6 +231,18 @@ function updateSupplier(){
 function deleteSupplier(){
     let suppCode = $('#supp_code').val();
 
+    let token = localStorage.getItem('user01');
+
+    // Check if token is available
+    if (!token) {
+        Swal.fire({
+            icon: 'error',
+            title: 'Authentication Error',
+            text: 'User not authenticated. Please log in.',
+        });
+        return;
+    }
+
     $.ajax({
         method:"DELETE",
         url:"http://localhost:8080/api/v1/suppliers/deleteSupplier/"+suppCode,
